Fix the navigation bar on smaller screens
The navigation bar was completely broken on smaller screens like on
tablets and phones. Part of the issue was due to the fixed and collapsed
capabilities of the nav bar. Those don't provide much value, so just
remove them completely. The navbar now scrolls along with the page, and
on small screen the nav bar just wraps to the next line and becomes
centered.