Change unsupported page to only list unsupported features in the current version
All the green checkmarks were pretty, but made it hard to actually
figure out what wasn't supported. So remove everything that is currently
supported. Also remove previous version support. That information isn't
all the important and is always available in the website source if we
need to figure it out. Also updated for the next release.

DAFFODIL-2039