Automate building/publishing of the website
Creating the website can be completely automated, and can be done in a
way that avoids environmental differences between developers systems.
This configured a github action to automatically build and publish the
github website everytime the master branch is update. A Gem file is used
to define dependencies and ensure consistency. The publish.sh script is
no longer needed, so it is deleted.