New asciidoc design notes on schema compiler space/speed issue.
There are 3 new notes:

* term-sharing-in-the-schema-compiler.adoc
* hidden-groups.adoc
* namespace-binding-minimization.adoc

They can be read on the Daffodil web site, or you can check-out the daffodil-site repository and view them directly. There is a PR for review comments on these asciidoc texts.

https://github.com/apache/incubator-daffodil-site/pull/17

These go along with a wiki document:

https://cwiki.apache.org/confluence/display/DAFFODIL/Analysis%3A+Schema+Compiler+Backpointers+and+Copying

Changes to the aboutAsciidoc.adoc file are just about the approach of using asciidoc
for this sort of design note.

Based on some preliminary reviews, added transition sections for
isHidden and namespace bindings notes.

Also fix bootstrap stylesheet with essentials for asciidoc pages.

DAFFODIL-1444