Only publish site updates from the master branch
We currently enable the publish action when commits are made directly to
the apache/daffodil-site repo. This prevents publishing pull requests
until they are merged. However, if we push a branch directly to
apache/daffodil-site for a pull request rather than a fork, then this
action is still triggered, and we will publish from the branch. We don't
want that. We only want to ever publish commits made to the master
branch of the apache/daffodil repo. This fixes that.