Use a generic name/email for website publish commit author
When committing website content in the Github publish workflow, we use
the github.event.head_commit.committer.name/email values, which should
be the name/email of the committer of the change. But sometimes it uses
the name/email associated with the account who merged the account (I
think), which is not necessarily the correct information. In my case,
its uses a personal email instead of an @apache.org email. Since this
doesn't information doesn't really matter, and since it's an automated
commit, switch to using dev@daffodil.apache.org and a generic Daffodil
name when committing changes to the asf-site branch.