Support release candidate download links
The recent changes to how we create download links was broken for
release candidates. To support this, this removes the "released" boolean
property and replaces it with a "release" property. This new property
should be set to "final" for released artifacts or set to the release
candidate name (e.g. "rc1") for release candidates. The download links
are modified to point to dist.apache.org and to include the "-rcX" in
the version name for release candidates.