Layer documentation added.
The Documentation pull-down at the top of the page now has an additional entry
for the extensions page, which has layer documentation, among other things.

Updated extensions page, albiet there are TBDs for several features.

The only changes to the site are this layer documentation, and the associated
improved javadoc/scaladoc for the layers.api classes.

This includes the 3.8.0 javadoc and scaladoc, which are linked from the
layers documentation.

The javadoc & scaladoc are copied here from the daf-2569-layer-doc branch
of the daffodil repository.

DAFFODIL-2569