Add workflow to manually update documentation from daffodil repositories
One of the steps the daffodil release candidate container did was update
the daffodil-site repository with documenation for the release candiate
(e.g. Java/Scala APIs, tutorials, VS Code documentation).

When we switch to a release workflow using GitHub action, we can no
longer do that--GitHub actions do not provide an easy way to commit to
other repositories like this daffodil-site repository.

To generate site documentation for release candidates, this adds a new
"Update Documentation" workflow to the site repository that should be
manually triggered via workflow_dispatch after the release candidate is
created. This workflow accepts the project and tag to build
documentation from, runs the appropriate commads to build the
documentation, and commits the updates to the site repo.

Note that due to GitHub actions limiations, pushing to the main branch
from this action does not trigger the CI/publish workflow that would
normally be triggered. So this also modifies the CI workflow so it can
be manually triggered. After the new "Update Documentation" workflow is
complete, we then must manually trigger the CI workflow to publish those
changes to the website.

So this means two new steps are added to a release workflow:

1. Manually trigger this workflow to build documentation
2. Manually trigger the publish workflow to build and publish the website

Although this does add a couple of extra steps to the release candidate
process, these two steps are just a couple clicks in the GitHub UI, and
the simplicity provided by switching releases to to GitHub actions
should more than makes up for it.

This also renames the 'main' workflow to 'build-publish' to make it more
clear that it publishes the website.

DAFFODIL-2791