Showcase pandoc pdf files and make script
Please run the site locally and look at the dfdl-extensions page and the layers page, and click the link to show the PDF files (near top of page).

This contains pandoc so a common markdown file
with some conditionalizations in it
can generate both web pages using
jekyll as usual, but also
pdf files from the same markdown.

Currently this requires you to cd to the site/_pandoc/ directory and type 'make' to run the makefile.
That will find all pages with 'pdf: true' in their header
and use pandoc to create PDFs for them.

The PDFs are placed in site/pdf, and are currently checked into the repo.

A drawback is you have to use a common markdown subset that works in jekyll and pandoc.

And you have to remember to 'make' the PDFs.

This is intended for a subset of our documenatation to use. Not every page has to have a PDF.

DAFFODIL-3044