Add apachecon and meetup powerpoint slides
Follow date-first naming convention so that newer/older material
can be distinguished even if files are downloaded and modified
data info is not preserved.

Contributes to DAFFODIL-3046 (Training).

These can be a source of other documentation also. E.g., some of the best
doc about the Daffodil streaming unparser is in the 2018 apachecon
talk.

DAFFODIL-3046