New page on best practices.
Created from content on the Daffodil Wiki "DFDL Schema Style Guide" page to start with.

Content revised/improved/simplified some.

More materials are to come. This just cuts over most of what was on the Wiki that is really for schema authors.

Directory created for various slides and other best practice material to live.

DAFFODIL-2998